version 1.0.0
The website, where links on different pages works and it has a form to create articles there. It was done with django, working with requests, models and most of thing were new for me, so it was useful.

P.S If you would like to test this project, just download archive, wich I will add here, open it in vs code, or different editor, change directory to myProject, and run local server with manage.py on your machine.